test(relay-mclp): consolidate and trim test suite
Replace 6 phase-split files (59 tests, ~350s) with one focused file
(15 tests, ~72s). Removed redundant, trivial, and slow tests.

Kept tests cover:
- Case I/O and candidate generation bounds
- Link geometry (ground elevation, ISL range/occultation)
- MCLP demand sampling, covered-samples computation, greedy marginal gain
- Scheduler degree-cap enforcement and interval compaction
- MILP bound rejection and auto-mode fallback
- Parallel propagation + link-cache equivalence vs sequential
- One end-to-end smoke test (solver + verifier on public case)
- Config-driven default grid candidate count

All 15 solver tests pass in ~72s.
All 5 experiment tests pass.
